+++ accounts-management/common/publickey.go
@@ -4,7 +4,7 @@ import (
 	"reflect"
 
 	"github.com/status-im/status-go/accounts-management/types"
-	"github.com/status-im/status-go/api/multiformat"
+	"github.com/status-im/status-go/pkg/multiformat"
 	"github.com/status-im/status-go/protocol/identity/emojihash"
 )
