+++ pkg/multiformat/utils.go
